Split API KBV routes into separate controller
Dependency injection means that the KBVService (and all its dependencies) is initialised for any controller that references it in the constructor.

By sharding the controllers and only depending what we need, we only initialise SoapClients when necessary and speed up every other page.

For ID-339 #patch
