+++ service-api/module/Application/src/Controller/IdentityController.php
@@ -4,21 +4,14 @@ declare(strict_types=1);
 
 namespace Application\Controller;
 
-use Application\Aws\Secrets\AwsSecret;
-use Application\Nino\ValidatorInterface;
 use Application\DrivingLicense\ValidatorInterface as LicenseValidatorInterface;
-use Application\Passport\ValidatorInterface as PassportValidator;
-use Application\KBV\KBVServiceInterface;
-use Application\Fixtures\DataWriteHandler;
 use Application\Fixtures\DataQueryHandler;
+use Application\Fixtures\DataWriteHandler;
 use Application\Model\Entity\CaseData;
 use Application\Model\Entity\Problem;
+use Application\Nino\ValidatorInterface;
+use Application\Passport\ValidatorInterface as PassportValidator;
 use Application\View\JsonModel;
-use Application\Yoti\Http\Exception\YotiException;
-use Application\Yoti\SessionConfig;
-use Application\Yoti\YotiServiceInterface;
-use DateTime;
-use Laminas\Cache\Storage\PluginManager;
 use Laminas\Form\Annotation\AttributeBuilder;
 use Laminas\Http\Response;
 use Laminas\Mvc\Controller\AbstractActionController;
@@ -39,7 +32,6 @@ class IdentityController extends AbstractActionController
         private readonly DataWriteHandler $dataHandler,
         private readonly LicenseValidatorInterface $licenseValidator,
         private readonly PassportValidator $passportService,
-        private readonly KBVServiceInterface $KBVService,
         private readonly LoggerInterface $logger
     ) {
     }
@@ -66,10 +58,12 @@ class IdentityController extends AbstractActionController
                 $this->dataHandler->insertUpdateData($caseData);
             } catch (\Exception $exception) {
                 $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
                 return new JsonModel(new Problem($exception->getMessage()));
             }
 
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
+
             return new JsonModel(['uuid' => $caseData->id]);
         }
 
@@ -88,6 +82,7 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
+
             return new JsonModel(new Problem('Missing uuid'));
         }
 
@@ -99,6 +94,7 @@ class IdentityController extends AbstractActionController
         }
 
         $this->getResponse()->setStatusCode(Response::STATUS_CODE_404);
+
         return new JsonModel(new Problem('Case not found'));
     }
 
@@ -120,12 +116,12 @@ class IdentityController extends AbstractActionController
         $data = [
             [
                 'lpa_ref' => 'PW PA M-XYXY-YAGA-35G3',
-                'donor_name' => 'Mary Anne Chapman'
+                'donor_name' => 'Mary Anne Chapman',
             ],
             [
                 'lpa_ref' => 'PW M-VGAS-OAGA-34G9',
-                'donor_name' => 'Mary Anne Chapman'
-            ]
+                'donor_name' => 'Mary Anne Chapman',
+            ],
         ];
 
         return new JsonModel($data);
@@ -138,7 +134,7 @@ class IdentityController extends AbstractActionController
         $ninoStatus = $this->ninoService->validateNINO($data['nino']);
 
         $response = [
-            'status' => $ninoStatus
+            'status' => $ninoStatus,
         ];
 
         $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
@@ -152,7 +148,7 @@ class IdentityController extends AbstractActionController
         $licenseStatus = $this->licenseValidator->validateDrivingLicense($data['dln']);
 
         $response = [
-            'status' => $licenseStatus
+            'status' => $licenseStatus,
         ];
         $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
 
@@ -165,86 +161,10 @@ class IdentityController extends AbstractActionController
         $passportStatus = $this->passportService->validatePassport(intval($data['passport']));
 
         $response = [
-            'status' => $passportStatus
+            'status' => $passportStatus,
         ];
         $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
 
-        return new JsonModel($response);
-    }
-
-    public function getKbvQuestionsAction(): JsonModel
-    {
-        $uuid = $this->params()->fromRoute('uuid');
-
-        if (! $uuid) {
-            $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
-            return new JsonModel(new Problem('Missing UUID'));
-        }
-
-        $case = $this->dataQueryHandler->getCaseByUUID($uuid);
-
-        if (is_null($case) || $case->documentComplete === false) {
-            $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
-            $response = [
-                "error" => "Document checks incomplete or unable to locate case"
-            ];
-            return new JsonModel($response);
-        }
-
-        $questionsWithoutAnswers = [];
-
-        $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
-
-        if (! is_null($case->kbvQuestions)) {
-            $questions = json_decode($case->kbvQuestions, true);
-
-            foreach ($questions as $number => $question) {
-                unset($question['answer']);
-                $questionsWithoutAnswers[$number] = $question;
-            }
-            //revisit formatting here, special character outputs
-            return new JsonModel($questionsWithoutAnswers);
-        } else {
-            $questions = $this->KBVService->fetchFormattedQuestions($uuid);
-
-            $this->dataHandler->updateCaseData(
-                $uuid,
-                'kbvQuestions',
-                'S',
-                json_encode($questions['formattedQuestions'])
-            );
-        }
-
-        return new JsonModel($questions['questionsWithoutAnswers']);
-    }
-
-    public function checkKbvAnswersAction(): JsonModel
-    {
-        $uuid = $this->params()->fromRoute('uuid');
-        $data = json_decode($this->getRequest()->getContent(), true);
-        $case = $this->dataQueryHandler->getCaseByUUID($uuid);
-
-        $result = 'pass';
-        $response = [];
-
-        if (! $uuid || is_null($case)) {
-            $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
-            return new JsonModel(new Problem("Missing UUID or unable to find case"));
-        }
-
-        $questions = json_decode($case->kbvQuestions, true);
-        //compare against all stored answers to ensure all answers passed
-        foreach ($questions as $key => $question) {
-            if (! isset($data['answers'][$key])) {
-                $result = 'fail';
-            } elseif ($data['answers'][$key] != $question['answer']) {
-                $result = 'fail';
-            }
-        }
-
-        $response['result'] = $result;
-
-        $this->getResponse()->setStatusCode(Response::STATUS_CODE_200);
         return new JsonModel($response);
     }
 
@@ -256,8 +176,10 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
+
             return new JsonModel(new Problem("Missing UUID"));
         }
+
         try {
             $this->dataHandler->updateCaseData(
                 $uuid,
@@ -267,6 +189,7 @@ class IdentityController extends AbstractActionController
             );
         } catch (\Exception $exception) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
 
@@ -284,8 +207,10 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
+
             return new JsonModel(new Problem('Missing UUID'));
         }
+
         try {
             $this->dataHandler->updateCaseData(
                 $uuid,
@@ -295,6 +220,7 @@ class IdentityController extends AbstractActionController
             );
         } catch (\Exception $exception) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
 
@@ -312,10 +238,11 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
+
             return new JsonModel(new Problem('Missing UUID'));
         }
         $counterServiceMap = [
-            "selectedPostOffice" => $data['selected_postoffice']
+            "selectedPostOffice" => $data['selected_postoffice'],
         ];
 
         try {
@@ -324,11 +251,12 @@ class IdentityController extends AbstractActionController
                 'counterService',
                 'M',
                 array_map(fn (mixed $v) => [
-                    'S' => $v
+                    'S' => $v,
                 ], $counterServiceMap),
             );
         } catch (\Exception $exception) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
 
@@ -348,6 +276,7 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_400);
+
             return new JsonModel(new Problem('Missing UUID'));
         }
 
@@ -363,11 +292,12 @@ class IdentityController extends AbstractActionController
                 'counterService',
                 'M',
                 array_map(fn (mixed $v) => [
-                    'S' => $v
+                    'S' => $v,
                 ], $counterServiceMap),
             );
         } catch (\Exception $exception) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
 
@@ -384,6 +314,7 @@ class IdentityController extends AbstractActionController
         $data = $this->dataQueryHandler->getCaseByUUID($uuid);
         $response = [];
         $status = Response::STATUS_CODE_200;
+
         try {
             /**
              * @psalm-suppress PossiblyNullPropertyFetch
@@ -407,6 +338,7 @@ class IdentityController extends AbstractActionController
         }
 
         $this->getResponse()->setStatusCode($status);
+
         return new JsonModel($response);
     }
 
@@ -417,6 +349,7 @@ class IdentityController extends AbstractActionController
         $data = $this->dataQueryHandler->getCaseByUUID($uuid);
         $response = [];
         $status = Response::STATUS_CODE_200;
+
         try {
             /**
              * @psalm-suppress PossiblyNullPropertyFetch
@@ -445,6 +378,7 @@ class IdentityController extends AbstractActionController
         }
 
         $this->getResponse()->setStatusCode($status);
+
         return new JsonModel($response);
     }
 
@@ -459,8 +393,9 @@ class IdentityController extends AbstractActionController
             $status = Response::STATUS_CODE_400;
             $this->getResponse()->setStatusCode($status);
             $response = [
-                "error" => "Missing UUID"
+                "error" => "Missing UUID",
             ];
+
             return new JsonModel($response);
         }
 
@@ -470,12 +405,13 @@ class IdentityController extends AbstractActionController
                 'alternateAddress',
                 'M',
                 array_map(fn (mixed $v) => [
-                    'S' => $v
+                    'S' => $v,
                 ], $data),
             );
         } catch (\Exception $exception) {
             $response['result'] = "Not Updated";
             $response['error'] = $exception->getMessage();
+
             return new JsonModel($response);
         }
 
@@ -496,8 +432,9 @@ class IdentityController extends AbstractActionController
             $status = Response::STATUS_CODE_400;
             $this->getResponse()->setStatusCode($status);
             $response = [
-                "error" => "Missing UUID"
+                "error" => "Missing UUID",
             ];
+
             return new JsonModel($response);
         }
 
@@ -511,6 +448,7 @@ class IdentityController extends AbstractActionController
         } catch (\Exception $exception) {
             $response['result'] = "Not Updated";
             $response['error'] = $exception->getMessage();
+
             return new JsonModel($response);
         }
 
@@ -532,8 +470,9 @@ class IdentityController extends AbstractActionController
             $status = Response::STATUS_CODE_400;
             $this->getResponse()->setStatusCode($status);
             $response = [
-                "error" => "Missing UUID"
+                "error" => "Missing UUID",
             ];
+
             return new JsonModel($response);
         }
 
@@ -541,8 +480,9 @@ class IdentityController extends AbstractActionController
             $status = Response::STATUS_CODE_400;
             $this->getResponse()->setStatusCode($status);
             $response = [
-                "error" => "Missing Date of Birth"
+                "error" => "Missing Date of Birth",
             ];
+
             return new JsonModel($response);
         }
 
@@ -556,6 +496,7 @@ class IdentityController extends AbstractActionController
         } catch (\Exception $exception) {
             $response['result'] = "Not Updated";
             $response['error'] = $exception->getMessage();
+
             return new JsonModel($response);
         }
 
@@ -574,6 +515,7 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem("Missing UUID"));
         }
 
@@ -583,12 +525,13 @@ class IdentityController extends AbstractActionController
                 'idMethodIncludingNation',
                 'M',
                 array_map(fn (mixed $v) => [
-                    'S' => $v
+                    'S' => $v,
                 ], $data),
             );
         } catch (\Exception $exception) {
             $this->logger->error($exception->getMessage());
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
 
@@ -611,6 +554,7 @@ class IdentityController extends AbstractActionController
 
         if (! $uuid) {
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem("Missing UUID"));
         }
 
@@ -620,12 +564,13 @@ class IdentityController extends AbstractActionController
                 'progressPage',
                 'M',
                 array_map(fn (mixed $v) => [
-                    'S' => $v
+                    'S' => $v,
                 ], $data),
             );
         } catch (\Exception $exception) {
             $this->logger->error($exception->getMessage());
             $this->getResponse()->setStatusCode(Response::STATUS_CODE_500);
+
             return new JsonModel(new Problem($exception->getMessage()));
         }
